requirements.txt
asgiref==3.3.1
certifi==2020.12.5
DAWG-Python==0.7.2
Django==3.1.7
django-debug-toolbar==3.2
djangorestframework==3.12.2
docopt==0.6.2
flake8==3.8.4
gunicorn==20.0.4
importlib-metadata==3.4.0
isort==5.7.0
mccabe==0.6.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pytz==2020.5
sentry-sdk==0.19.5
sorl-thumbnail==12.7.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
zipp==3.4.0
